Move next statement code into db & get rid of next statement altogether
